[feat] Replace translate action
Translate action was replaced to profile like caction cell.
