+++ Client/src/pages/Profile/components/TranslateCell/TranslateCell.scss
@@ -1,0 +1,8 @@
+img.tc__icon {
+	height: 2.4vh;
+	border-radius: 4px;
+
+	&#arrow {
+		filter: brightness(0) saturate(100%) invert(100%) sepia(0%) saturate(0%) hue-rotate(93deg) brightness(103%) contrast(103%);
+	}
+}
